requirements.txt

mysql-connector        2.2.9
mysql-connector-python 8.1.0
pip                    23.2.1
requests               2.31.0
